TDD in Unity 6 지속적으로 변수이름과 구조에 대한 리펙토링을 진행하면서 테스트코드를 추가함. https://youtu.be/x7gxKmhcOfw

+++ Assets/Scripts/TDD in Unity/Infrastructure/ImageBuilder.cs
@@ -5,15 +5,15 @@ namespace TDD_in_Unity.Infrastructure
 {
 	public class ImageBuilder : TestDataBuilder<Image>
 	{
-		private int _fillAmount;
+		private float _fillAmount;
 
-		public ImageBuilder(int fillAmount) {
+		public ImageBuilder(float fillAmount) {
 			_fillAmount = fillAmount;
 		}
 
 		public ImageBuilder() : this(0) { }
 
-		public ImageBuilder WithFillAmount(int fillAmount) {
+		public ImageBuilder WithFillAmount(float fillAmount) {
 			_fillAmount = fillAmount;
 			return this;
 		}
